Fix #296 #299 - Fix issue where `encryptECIES` would sometimes result in a `TypeError: Invalid Hex String`
This issue was caused when the length of `sharedSecret.toString('hex')` was not 64.

To work around this issue a new recursive function `getEncryptionKeys` was created that will
return an object with `sharedSecret` and `ephemeralPK` to be used by `encryptECIES`. The function will only
return a value if the converted hex string for `sharedSecret` is 64 characters long. The function will run again
otherwise.